Introduce quality control callbacks for AST tests
The qcfunc is intended to be a function that takes the structured numpy
array of a FakeTable, and the index of the image, and return a boolean
array indicated if that artificial star would be accepted according to
quality control criteria (sharpness, roundness, etc) set by the user.
This allow completeness estimates to include the user's filtering.

The qc callback function is directly applied in recovered_in_image().

It is also enabled for use from ASTReduce and
FakeReader.export_for_starfish().