Create PhotTable subclass of astropy Table
The PhotTable.read_phot class method runs the parser to create a numpy
record array that then seeds the table. Since PhotTable is an astropy
Table, all of its methods re available.

We use the metadata to record the number of images, as well as the
names/bandpasses of photometered images. This metadata is available PhotTable
properties image_names and _bands and n_images.

Use the PhotTable.image_col method for reading image-specific columns.
Since the photometery of multiple image is provided in a flat table,
columns such as 'mag' have the image index appended to their name. This
image_col method can resolve and read from the correct column given the
index, image name, or bandpass.