Internally encode Table metadata in JSON
If a metadata item is set as a non str/float/int style, such as a list,
then it needs to be encoded as a string. I've chosen to do this with
json encoding.

The metadata is automatically encoded during the read_phot class
methods, and are un-encoded as necessary via the properties. Note that
metadata fields not covered by the properties will not be unencoded.
This could certainly be fixed in the future with some dynamic
properties.